ai_consumer_insights/requirements.txt
streamlit==1.28.1
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.2
bertopic==0.15.0
sentence-transformers==2.2.2
spacy==3.7.2
prophet==1.1.4
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
